[BE] refactor: ApiDocSnippets 패키지명 및 클래스명 변경 #205
